Set up front end with vue3 + typescript authentication and state management bug fixes update read me

+++ src/DTO/ReviewRequestDTO.php
@@ -7,7 +7,7 @@ class ReviewRequestDTO
 {
     #[Assert\NotBlank(message: "Rating should not be blank.")]
     #[Assert\Range(min: 1, max: 5, notInRangeMessage: "Rating must be between 1 and 5.")]
-    public int $rating;
+    public float $rating;
 
     #[Assert\NotBlank(message: "Comment should not be blank.")]
     #[Assert\Type(type: "string", message: "Comment must be a valid text.")]
